[LOVE-11] Rename user dto file
