+++ src/objects/dtos/game/players/a-player.dto.ts
@@ -1,4 +1,4 @@
-import { UserDTO } from "../../users/userDTO";
+import { UserDTO } from "../../users/user.dto";
 
 export abstract class APlayerDTO {
   // *****************************************************************************************************************
